requirements.txt
streamlit
easyocr
pillow
numpy
requests
opencv-python-headless
